refactor out DataProvider
